+++ njunmt/data/dataset.py
@@ -18,12 +18,6 @@ from __future__ import print_function
 
 import six
 
-from tensorflow.python.ops import array_ops
-from tensorflow.python.ops import math_ops
-from tensorflow.python.framework import ops
-from tensorflow.python.framework import dtypes
-
-from njunmt.utils.constants import Constants
 from njunmt.utils.misc import get_labels_files
 
 
@@ -58,50 +52,6 @@ class Dataset(object):
             self._eval_labels_file = get_labels_files(eval_labels_file)
         else:
             self._eval_labels_file = None
-        self._input_fields = Dataset._make_input_fields()
-
-    @staticmethod
-    def _make_input_fields():
-        """ Creates tf placeholders and add input data status
-              to tf collections for displaying.
-
-        Returns: A dictionary of placeholders.
-
-        """
-        feature_ids = array_ops.placeholder(dtypes.int32, shape=(None, None), name=Constants.FEATURE_IDS_NAME)
-        feature_length = array_ops.placeholder(dtypes.int32, shape=(None,), name=Constants.FEATURE_LENGTH_NAME)
-        label_ids = array_ops.placeholder(dtypes.int32, shape=(None, None), name=Constants.LABEL_IDS_NAME)
-        label_length = array_ops.placeholder(dtypes.int32, shape=(None,), name=Constants.LABEL_LENGTH_NAME)
-
-        feature_nonpadding_tokens_num = math_ops.reduce_sum(feature_length)
-        feature_shape = array_ops.shape(feature_ids)
-        feature_total_tokens_num = feature_shape[0] * feature_shape[1]
-        label_nonpadding_tokens_num = math_ops.reduce_sum(label_length)
-        label_shape = array_ops.shape(label_ids)
-        label_total_tokens_num = label_shape[0] * label_shape[1]
-        ops.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/feature_nonpadding_tokens_num")
-        ops.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME, feature_nonpadding_tokens_num)
-        ops.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/feature_nonpadding_ratio")
-        ops.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME,
-                              math_ops.to_float(feature_nonpadding_tokens_num)
-                              / math_ops.to_float(feature_total_tokens_num))
-        ops.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/label_nonpadding_tokens_num")
-        ops.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME, label_nonpadding_tokens_num)
-        ops.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/label_nonpadding_ratio")
-        ops.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME,
-                              math_ops.to_float(label_nonpadding_tokens_num)
-                              / math_ops.to_float(label_total_tokens_num))
-        return {Constants.FEATURE_IDS_NAME: feature_ids,
-                Constants.FEATURE_LENGTH_NAME: feature_length,
-                Constants.LABEL_IDS_NAME: label_ids,
-                Constants.LABEL_LENGTH_NAME: label_length}
-
-    @property
-    def input_fields(self):
-        """ Returns the dictionary of placeholders. """
-        if self._input_fields is None:
-            self._input_fields = Dataset._make_input_fields()
-        return self._input_fields
 
     @property
     def vocab_source(self):