+++ njunmt/models/sequence_to_sequence.py
@@ -33,7 +33,6 @@ from njunmt.utils.constants import ModeKeys
 from njunmt.utils.beam_search import process_beam_predictions
 from njunmt.utils.misc import set_fflayers_layer_norm
 
-
 # import all bridges
 BRIDGE_CLSS = [
     x for x in bridges.__dict__.values()
@@ -46,6 +45,8 @@ for bri in BRIDGE_CLSS:
 @six.add_metaclass(ABCMeta)
 class SequenceToSequence(Configurable):
     """ Base Sequence-to-Sequence Class """
+
+    __MODEL_COUNTER = 0
 
     def __init__(self,
                  params,
@@ -66,12 +67,72 @@ class SequenceToSequence(Configurable):
             verbose: Print model parameters if set True.
         """
         super(SequenceToSequence, self).__init__(
-            params=params, mode=mode, verbose=False,
-            name=name)
+            params=params, mode=mode, verbose=False, name=name)
         self._vocab_source = vocab_source
         self._vocab_target = vocab_target
         self._verbose = verbose
+        self._input_fields = SequenceToSequence.__create_input_fields(self.mode)
         set_fflayers_layer_norm(self.params["fflayers.layer_norm"])
+
+    @staticmethod
+    def __create_input_fields(mode):
+        """ Creates tf placeholders and add input data status to tf
+          collections for displaying.
+
+        Returns: A dictionary of placeholders.
+        """
+        SequenceToSequence.__MODEL_COUNTER += 1
+        inp = dict()
+        feature_ids = tf.placeholder(
+            tf.int32, shape=(None, None),
+            name="{}_{}".format(Constants.FEATURE_IDS_NAME, SequenceToSequence.__MODEL_COUNTER - 1))
+        feature_length = tf.placeholder(
+            tf.int32, shape=(None,),
+            name="{}_{}".format(Constants.FEATURE_LENGTH_NAME, SequenceToSequence.__MODEL_COUNTER - 1))
+        inp[Constants.FEATURE_IDS_NAME] = feature_ids
+        inp[Constants.FEATURE_LENGTH_NAME] = feature_length
+        if mode == ModeKeys.INFER:
+            return inp
+
+        label_ids = tf.placeholder(
+            tf.int32, shape=(None, None),
+            name="{}_{}".format(Constants.LABEL_IDS_NAME, SequenceToSequence.__MODEL_COUNTER - 1))
+        label_length = tf.placeholder(
+            tf.int32, shape=(None,),
+            name="{}_{}".format(Constants.LABEL_LENGTH_NAME, SequenceToSequence.__MODEL_COUNTER - 1))
+        inp[Constants.LABEL_IDS_NAME] = label_ids
+        inp[Constants.LABEL_LENGTH_NAME] = label_length
+        if mode == ModeKeys.EVAL:
+            return inp
+
+        assert mode == ModeKeys.TRAIN
+        feature_nonpadding_tokens_num = tf.reduce_sum(feature_length)
+        feature_shape = tf.shape(feature_ids)
+        feature_total_tokens_num = feature_shape[0] * feature_shape[1]
+        tf.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/feature_nonpadding_tokens_num")
+        tf.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME, feature_nonpadding_tokens_num)
+        tf.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/feature_nonpadding_ratio")
+        tf.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME,
+                             tf.to_float(feature_nonpadding_tokens_num)
+                             / tf.to_float(feature_total_tokens_num))
+
+        label_nonpadding_tokens_num = tf.reduce_sum(label_length)
+        label_shape = tf.shape(label_ids)
+        label_total_tokens_num = label_shape[0] * label_shape[1]
+
+        tf.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/label_nonpadding_tokens_num")
+        tf.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME, label_nonpadding_tokens_num)
+        tf.add_to_collection(Constants.DISPLAY_KEY_COLLECTION_NAME, "input_stats/label_nonpadding_ratio")
+        tf.add_to_collection(Constants.DISPLAY_VALUE_COLLECTION_NAME,
+                             tf.to_float(label_nonpadding_tokens_num)
+                             / tf.to_float(label_total_tokens_num))
+        return inp
+
+    @property
+    def input_fields(self):
+        if self._input_fields is None:
+            self._input_fields = SequenceToSequence.__create_input_fields(self.mode)
+        return self._input_fields
 
     def _create_modalities(self):
         """ Creates source and target modalities.
@@ -136,14 +197,11 @@ class SequenceToSequence(Configurable):
         else:
             raise ValueError("Unrecognized initializer: {}".format(self.params["initializer"]))
 
-    def build(self, input_fields):
+    def build(self):
         """ Builds the sequence-to-sequence model.
 
         This function calls many inner functions to build each component
         of the model.
-
-        Args:
-            input_fields: A dictionary of placeholders.
 
         Returns: Model output. See _pack_output() for more details.
         """
@@ -152,8 +210,7 @@ class SequenceToSequence(Configurable):
             encoder = self._create_encoder()
             encoder_output = self._encode(
                 encoder=encoder,
-                input_modality=input_modality,
-                input_fields=input_fields)
+                input_modality=input_modality)
 
             encdec_bridge = self._create_bridge(encoder_output)
             decoder = self._create_decoder()
@@ -161,12 +218,11 @@ class SequenceToSequence(Configurable):
                 decoder=decoder,
                 encdec_bridge=encdec_bridge,
                 encoder_output=encoder_output,
-                target_modality=target_modality,
-                input_fileds=input_fields)
+                target_modality=target_modality)
 
             final_outputs = self._pack_output(
                 encoder_output, decoder_output, decoding_res,
-                target_modality, **input_fields)
+                target_modality)
         return final_outputs
 
     def _compute_loss(self, logits, label_ids, label_length, target_modality):
@@ -186,7 +242,7 @@ class SequenceToSequence(Configurable):
             return loss
 
     def _decode(self, decoder, encdec_bridge, encoder_output,
-                target_modality, input_fileds):
+                target_modality):
         """ Builds helper and calls decoder's `decode` method.
 
         Args:
@@ -195,22 +251,21 @@ class SequenceToSequence(Configurable):
             encoder_output: An instance of `collections.namedtuple`
               from `Encoder.encode()`.
             target_modality: An instance of `Modality`.
-            input_fields: A dictionary of placeholders.
 
         Returns: The results of decoding. For more details, see
           `Decoder.decode()`.
         """
         if self.mode == ModeKeys.TRAIN \
                 or self.mode == ModeKeys.EVAL:
-            label_ids = input_fileds[Constants.LABEL_IDS_NAME]
-            label_length = input_fileds[Constants.LABEL_LENGTH_NAME]
+            label_ids = self.input_fields[Constants.LABEL_IDS_NAME]
+            label_length = self.input_fields[Constants.LABEL_LENGTH_NAME]
             helper = feedback.TrainingFeedback(
                 vocab=self._vocab_target, label_ids=label_ids, label_length=label_length)
 
         else:  # self.mode == tf.contrib.learn.ModeKeys.INFER
             helper = feedback.BeamFeedback(
                 vocab=self._vocab_target,
-                batch_size=tf.shape(input_fileds[Constants.FEATURE_IDS_NAME])[0],
+                batch_size=tf.shape(self.input_fields[Constants.FEATURE_IDS_NAME])[0],
                 maximum_labels_length=self.params["inference.maximum_labels_length"],
                 beam_size=self.params["inference.beam_size"],
                 alpha=self.params["inference.length_penalty"])
@@ -219,19 +274,18 @@ class SequenceToSequence(Configurable):
             beam_size=self.params["inference.beam_size"])
         return decoder_output, decoding_res
 
-    def _encode(self, encoder, input_modality, input_fields):
+    def _encode(self, encoder, input_modality):
         """ Calls encoder's encode method.
 
         Args:
             encoder: An instance of `Encoder`.
             input_modality: An instance of `Modality`.
-            input_fields: A dictionary of placeholders.
 
         Returns: The results of encoding, an instance of `collections.namedtuple`
           from `Encoder.encode()`.
         """
-        feature_ids = input_fields[Constants.FEATURE_IDS_NAME]
-        feature_length = input_fields[Constants.FEATURE_LENGTH_NAME]
+        feature_ids = self.input_fields[Constants.FEATURE_IDS_NAME]
+        feature_length = self.input_fields[Constants.FEATURE_LENGTH_NAME]
 
         if self.params["source.reverse"]:
             feature_ids = tf.reverse_sequence(
@@ -313,8 +367,8 @@ class SequenceToSequence(Configurable):
         if self.mode == ModeKeys.TRAIN or self.mode == ModeKeys.EVAL:
             loss = self._compute_loss(
                 logits=decoding_result,  # [timesteps, batch_size, dim]
-                label_ids=kwargs[Constants.LABEL_IDS_NAME],
-                label_length=kwargs[Constants.LABEL_LENGTH_NAME],
+                label_ids=self.input_fields[Constants.LABEL_IDS_NAME],
+                label_length=self.input_fields[Constants.LABEL_LENGTH_NAME],
                 target_modality=target_modality)
         if self.mode == ModeKeys.TRAIN:
             return loss
@@ -345,5 +399,5 @@ class SequenceToSequence(Configurable):
             beam_size=self.params["inference.beam_size"],
             alpha=self.params["inference.length_penalty"])
         predict_out["attentions"] = attentions
-        predict_out["source"] = kwargs[Constants.FEATURE_IDS_NAME]
+        predict_out["source"] = self.input_fields[Constants.FEATURE_IDS_NAME]
         return predict_out